git-svn-id: http://svn.wso2.org/repos/wso2/trunk/wsf@75743 a5903396-d722-0410-b921-86c7d4935375

+++ cpp/include/OMDataHandler.h
@@ -111,7 +111,7 @@ namespace wso2wsf
          * @param output_stream_length size of retrieved stream.
          * @return true if operation was successful, false otherwise.
          */
-        WSF_EXTERN bool WSF_CALL read(axis2_byte_t ** output_stream, int * output_stream_length);
+        WSF_EXTERN bool WSF_CALL read(axis2_byte_t ** output_stream, size_t * output_stream_length);
 
         /**
          * Method to write to a stream.
